web_rebuild/backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic-settings>=2.1.0
aiomysql>=0.2.0
mysql-connector-python>=8.0.0
httpx>=0.26.0
akshare>=1.12.0
openai>=1.0.0
